+++ R/utils_expand_data_to_grid.R
@@ -2,7 +2,7 @@
 #' @importFrom sjstats pred_vars typical_value
 #' @importFrom sjmisc to_value to_factor
 #' @importFrom stats terms
-#' @importFrom purrr map map_lgl
+#' @importFrom purrr map map_lgl map_df
 get_expanded_data <- function(model, mf, terms, typ.fun) {
   # special handling for coxph
   if (inherits(model, "coxph")) mf <- dplyr::select(mf, -1)
@@ -113,27 +113,18 @@ get_sliced_data <- function(fitfram, terms) {
 
 
 get_cleaned_varnames <- function(x) {
+  # for gam-smoothers/loess, remove s()- and lo()-function in column name
+  # for survival, remove strata()
+  pattern <- c("log", "s", "lo", "bs", "poly", "strata")
+
   # do we have a "log()" pattern here? if yes, get capture region
   # which matches the "cleaned" variable name
-  for (i in 1:length(x)) {
-    pos <- regexpr(pattern = "log(.*)", text = x[i], perl = TRUE)
-    start <- attr(pos, "capture.start", exact = TRUE)
-    len <- attr(pos, "capture.length", exact = TRUE)
-
-    # return substring or original variable
-    if (start != -1 && len != -1)
-      x[i] <- substr(x[i], start = start + 1, stop = start + len - 2)
-    else
-      x[i]
-
-    # for gam-smoothers/loess, remove s()- and lo()-function in column name
-    x[i] <- unique(sub("^s\\(([^,)]*).*", "\\1", x[i]))
-    x[i] <- unique(sub("^lo\\(([^,)]*).*", "\\1", x[i]))
-
-    # for survival, remove strata()
-    x[i] <- unique(sub("^strata\\(([^,)]*).*", "\\1", x[i]))
-  }
-
-  x
+  purrr::map_chr(1:length(x), function(i) {
+    for (j in 1:length(pattern)) {
+      p <- paste0("^", pattern[j], "\\(([^,)]*).*")
+      x[i] <- unique(sub(p, "\\1", x[i]))
+    }
+    x[i]
+  })
 }